Rename termsConditionsScreen.ts to TermsConditionsScreen.ts
